Clean ui theming violations
