+++ frontend/src/embedding-sdk-bundle/types/theme.ts
@@ -1,8 +1,8 @@
 export type {
   MetabaseEmbeddingColorKeyV2,
-  MetabaseEmbeddingTheme,
   MetabaseEmbeddingThemeV2,
-} from "metabase/embedding-sdk/theme";
+} from "metabase/ui/colors/types";
+export type { MetabaseEmbeddingTheme } from "metabase/embedding-sdk/theme";
 
 export type {
   ChartColorV2,
