+++ frontend/src/metabase/data-grid/hooks/use-measure-column-widths.tsx
@@ -8,12 +8,10 @@ import type { Root } from "react-dom/client";
 
 import type { ColumnOptions, DataGridTheme } from "metabase/data-grid/types";
 import { pickRowsToMeasure } from "metabase/data-grid/utils/column-sizing";
-import {
-  MeasurementProviders,
-  createMeasurementContainer,
-} from "metabase/data-grid/utils/measure-utils";
+import { createMeasurementContainer } from "metabase/lib/measure-container";
 import { renderRoot, unmountRoot } from "metabase/lib/react-compat";
 import { isNotNull } from "metabase/lib/types";
+import { MeasurementProviders } from "metabase/ui/components/theme/MeasurementProviders";
 
 import { DataGridThemeProvider } from "./use-table-theme";
