+++ frontend/src/metabase/embedding-sdk/theme/index.ts
@@ -1,6 +1,5 @@
 export * from "./default-component-theme";
 export * from "./MetabaseTheme";
-export * from "./MetabaseThemeV2";
 export * from "./theme-version";
 
 export * from "./define-metabase-theme";
