+++ frontend/src/metabase/lib/measure-container.ts
@@ -1,15 +1,12 @@
-import type { ReactNode } from "react";
 import type { Root } from "react-dom/client";
 
 import { unmountRoot } from "metabase/lib/react-compat";
-import { EmotionCacheProvider } from "metabase/styled-components/components/EmotionCacheProvider";
-import { ThemeProvider } from "metabase/ui";
-
-import { DEFAULT_FONT_SIZE } from "../constants";
 
 interface MeasurementContainerOptions {
   fontSize?: string;
 }
+
+const DEFAULT_FONT_SIZE = "12.5px";
 
 export function createMeasurementContainer(
   options: MeasurementContainerOptions = {},
@@ -38,16 +35,4 @@ export function removeMeasurementContainer(
     unmountRoot(tree, container);
     document.body.removeChild(container);
   }, 0);
-}
-
-interface MeasurementProvidersProps {
-  children: ReactNode;
-}
-
-export function MeasurementProviders({ children }: MeasurementProvidersProps) {
-  return (
-    <EmotionCacheProvider>
-      <ThemeProvider>{children}</ThemeProvider>
-    </EmotionCacheProvider>
-  );
 }
