--- frontend/src/metabase/styled-components/components/EmotionCacheProvider/index.ts
@@ -1,1 +1,0 @@
-export * from "./EmotionCacheProvider";
