+++ frontend/src/metabase/ui/components/theme/EmotionCacheProvider/index.ts
@@ -1,0 +1,1 @@
+export { EmotionCacheProvider } from "./EmotionCacheProvider";
