+++ frontend/src/metabase/ui/components/theme/MeasurementProviders.tsx
@@ -1,0 +1,16 @@
+import type { ReactNode } from "react";
+
+import { EmotionCacheProvider } from "./EmotionCacheProvider";
+import { ThemeProvider } from "./ThemeProvider";
+
+interface MeasurementProvidersProps {
+  children: ReactNode;
+}
+
+export function MeasurementProviders({ children }: MeasurementProvidersProps) {
+  return (
+    <EmotionCacheProvider>
+      <ThemeProvider>{children}</ThemeProvider>
+    </EmotionCacheProvider>
+  );
+}
